Particles.

+++ src/pocketmine/network/protocol/LevelEventPacket.php
@@ -45,9 +45,9 @@ class LevelEventPacket extends DataPacket{
 	public function encode(){
 		$this->reset();
 		$this->putShort($this->evid);
-		$this->putInt($this->x);
-		$this->putByte($this->y);
-		$this->putInt($this->z);
+		$this->putFloat($this->x);
+		$this->putFloat($this->y);
+		$this->putFloat($this->z);
 		$this->putInt($this->data);
 	}
